+++ frontend/cruip-landing/theme.config.jsx
@@ -1,6 +1,18 @@
 export default {
-  logo: <>从txt到mp4</>,
-  project: { link: 'https://github.com/Golenspade/MyWebDrive' },
-  docsRepositoryBase: 'https://github.com/Golenspade/MyWebDrive/tree/main/frontend/cruip-landing/content',
+  logo: <span>从txt到mp4 文档</span>,
+  project: {
+    link: 'https://github.com/Golenspade/MyWebDrive',
+  },
+  docsRepositoryBase: 'https://github.com/Golenspade/MyWebDrive/tree/main/frontend/cruip-landing/pages',
+  useNextSeoProps() {
+    return {
+      titleTemplate: '%s – 从txt到mp4'
+    }
+  },
+  footer: {
+    text: <span>
+      MIT {new Date().getFullYear()} © <a href="https://mygoavemujica.top" target="_blank">MyGO Studio</a>.
+    </span>
+  }
 };